+++ docs/changelog.md
@@ -22,6 +22,8 @@
 - [#237](https://github.com/turicas/rows/issues/237) Implement `rows schema`
   (generates schema in text, SQL and Django models).
 - [#243](https://github.com/turicas/rows/issues/243) Change license to LGPL3.0
+- [#225](https://github.com/turicas/rows/issues/225) Create export callbacks
+  (CSV and SQLite plugins)
 
 ### Bug Fixes